Fixing serialization of boolean bug. Updating source code with new guides of Xcode 5.1. Migrating to XCTest unit testing

+++ Framework/ROADWebService/ROADWebService/Attributes/RFWebServiceCall.m
@@ -42,7 +42,7 @@
         _overrideGlobalSuccessCodes = YES;
         _method = @"GET";
         _prototypeClass = nil;
-        _postParameter = NSNotFound;
+        _postParameter = (int)NSNotFound;
     }
 
     return self;
